# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/books": {
            "get": {
                "description": "Ler e restorna os dados dos livros com categorias e autores, e filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Ler os livros",
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.GenericResponse-models_Books"
                        }
                    },
                    "404": {
                        "description": "Dados não existentes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo livro com os dados fornecidos, incluindo título, descrição, conteúdo, autor, categorias e imagem.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Cria um novo livro",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Título do Livro",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descrição do Livro",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Conteúdo do Livro",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID do Autor",
                        "name": "author_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Categorias do Livro",
                        "name": "categories",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagem do Livro",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.ResponseGeneric-models_Books"
                        }
                    },
                    "400": {
                        "description": "Erro nos dados fornecidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/books/{id}": {
            "put": {
                "description": "Atualiza um livro fornecido pelo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Atualiza um novo livro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do livro",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.Response"
                        }
                    },
                    "400": {
                        "description": "Requisição nao execultada",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Apaga um livro fornecido pelo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livros"
                ],
                "summary": "Apaga um novo livro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do livro",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dados não existentes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/categories": {
            "get": {
                "description": "Ler e restorna os dados das categorias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Ler as categorias",
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.GenericResponse-models_Categories"
                        }
                    },
                    "404": {
                        "description": "Dados não existentes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma categoria com os dados fornecidos, nome",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Cria uma categoria",
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.ResponseGeneric-models_Categories"
                        }
                    },
                    "400": {
                        "description": "Requisição nao execultada",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}": {
            "delete": {
                "description": "Apaga uma categoria fornecido pelo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categorias"
                ],
                "summary": "Apaga uma categoria livro",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do livro",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Dados não existentes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Faz o login do usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Faz o login do usuario",
                "responses": {
                    "200": {
                        "description": "Usuario criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.Response"
                        }
                    },
                    "400": {
                        "description": "Erro nos dados fornecidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/public/books": {
            "get": {
                "description": "Ler e restorna os dados dos usuarios com  filtros",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Ler os Usuarios",
                "responses": {
                    "200": {
                        "description": "Execultada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.GenericResponse-models_Users"
                        }
                    },
                    "404": {
                        "description": "Dados não existentes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Cria um novo usuario com os dados fornecidos, name, email, senha, username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Cria um novo Usuario",
                "responses": {
                    "201": {
                        "description": "Usuario criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.ResponseGeneric-models_Users"
                        }
                    },
                    "400": {
                        "description": "Erro nos dados fornecidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/users/{id}": {
            "put": {
                "description": "Atualiza usuarios com os dados fornecidos, name, bio, foto",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Atualiza usuarios",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do usuario",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descrição do usuario",
                        "name": "bio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Foto do usuario",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/validacao.ResponseGeneric-models_Users"
                        }
                    },
                    "400": {
                        "description": "Erro nos dados fornecidos",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Authors": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                }
            }
        },
        "models.Books": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Authors"
                    }
                },
                "author_id": {
                    "type": "integer"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Categories"
                    }
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Categories": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "validacao.GenericResponse-models_Books": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Books"
                    }
                }
            }
        },
        "validacao.GenericResponse-models_Categories": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Categories"
                    }
                }
            }
        },
        "validacao.GenericResponse-models_Users": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Users"
                    }
                }
            }
        },
        "validacao.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "validacao.ResponseGeneric-models_Books": {
            "type": "object",
            "properties": {
                "items": {
                    "$ref": "#/definitions/models.Books"
                }
            }
        },
        "validacao.ResponseGeneric-models_Categories": {
            "type": "object",
            "properties": {
                "items": {
                    "$ref": "#/definitions/models.Categories"
                }
            }
        },
        "validacao.ResponseGeneric-models_Users": {
            "type": "object",
            "properties": {
                "items": {
                    "$ref": "#/definitions/models.Users"
                }
            }
        }
    }
}